fix: Reset factory model selections
Reset the model editor when the selected provider or credential source
changes. Split the agent card components from the models page so each
module stays focused.
